feat(apps): more service implementations, cleanup, and tests
