+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Speckle Server",
+  "name": "speckle-server",
   "version": "2.0.0",
   "description": "",
   "homepage": "https://github.com/specklesystems/server",
